GridSuite migration
