+++ .github/workflows/maven.yml
@@ -23,8 +23,8 @@ jobs:
         run: >
           mvn --batch-mode -Pjacoco package sonar:sonar
           -Dsonar.host.url=https://sonarcloud.io
-          -Dsonar.organization=powsybl-ci-github
-          -Dsonar.projectKey=com.rte-france.powsybl:powsybl-odre-server
+          -Dsonar.organization=gridsuite
+          -Dsonar.projectKey=org.gridsuite:odre-server
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
