+++ README.md
@@ -1,6 +1,5 @@
 # [Open Data Reseaux Energies](https://opendata.reseaux-energies.fr/pages/accueil/) geographical data import service
 
-[![Actions Status](https://github.com/rte-france/powsybl-odre-server/workflows/CI/badge.svg)](https://github.com/rte-france/powsybl-odre-server/actions)
-[![Coverage Status](https://sonarcloud.io/api/project_badges/measure?project=com.rte-france.powsybl%3Apowsybl-odre-server&metric=coverage)](https://sonarcloud.io/component_measures?id=com.rte-france.powsybl%3Apowsybl-odre-server&metric=coverage)
+[![Actions Status](https://github.com/gridsuite/odre-server/workflows/CI/badge.svg)](https://github.com/gridsuite/odre-server/actions)
+[![Coverage Status](https://sonarcloud.io/api/project_badges/measure?project=com.gridsuite%3Aodre-server&metric=coverage)](https://sonarcloud.io/component_measures?id=org.gridsuite%3Aodre-server&metric=coverage)
 [![MPL-2.0 License](https://img.shields.io/badge/license-MPL_2.0-blue.svg)](https://www.mozilla.org/en-US/MPL/2.0/)
-[![Join the community on Spectrum](https://withspectrum.github.io/badge/badge.svg)](https://spectrum.chat/powsybl)
