--- src/main/java/com/rte_france/powsybl/odre/server/client/OdreClient.java
@@ -1,27 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.rte_france.powsybl.odre.server.client;
-
-import com.rte_france.powsybl.odre.server.dto.LineGeoData;
-import com.rte_france.powsybl.odre.server.dto.SubstationGeoData;
-
-import java.nio.file.Path;
-import java.util.List;
-
-/**
- * @author Chamseddine Benhamed <chamseddine.benhamed at rte-france.com>
- */
-public interface OdreClient {
-
-    List<SubstationGeoData> getSubstations();
-
-    List<SubstationGeoData> getSubstations(Path path);
-
-    List<LineGeoData> getLines();
-
-    List<LineGeoData> getLines(Path aerialLinesFilePath, Path undergroundLinesFilePath);
-}
