--- src/main/java/com/rte_france/powsybl/odre/server/dto/Coordinate.java
@@ -1,23 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.rte_france.powsybl.odre.server.dto;
-
-import lombok.*;
-
-/**
- * @author Chamseddine Benhamed <chamseddine.benhamed at rte-france.com>
- */
-@AllArgsConstructor
-@Getter
-@ToString
-@EqualsAndHashCode
-public class Coordinate {
-
-    private  double lat;
-
-    private  double lon;
-}
