--- src/main/java/com/rte_france/powsybl/odre/server/services/OdreService.java
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.rte_france.powsybl.odre.server.services;
-
-/**
- * @author Chamseddine Benhamed <chamseddine.benhamed at rte-france.com>
- */
-public interface OdreService {
-
-    void pushSubstations();
-
-    void pushLines();
-}
