+++ src/main/java/org/gridsuite/odre/server/OdreApplication.java
@@ -1,0 +1,21 @@
+/**
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.gridsuite.odre.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author Chamseddine Benhamed <chamseddine.benhamed at rte-france.com>
+ */
+@SpringBootApplication
+public class OdreApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(OdreApplication.class, args);
+    }
+}
