+++ src/main/java/org/gridsuite/odre/server/OdreSwaggerConfig.java
@@ -1,0 +1,49 @@
+/**
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.gridsuite.odre.server;
+
+import com.google.common.base.Predicate;
+import com.google.common.base.Predicates;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+/**
+ * @author Chamseddine Benhamed <chamseddine.benhamed at rte-france.com>
+ */
+@Configuration
+@EnableSwagger2
+public class OdreSwaggerConfig {
+    @Bean
+    public Docket produceApi() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .apiInfo(apiInfo())
+                .select()
+                .apis(RequestHandlerSelectors.basePackage(OdreController.class.getPackage().getName()))
+                .paths(paths())
+                .build();
+    }
+
+    private ApiInfo apiInfo() {
+        return new ApiInfoBuilder()
+                .title("Open Data Reseaux Energies geo data import API")
+                .description("This is the documentation of Open Data Reseaux Energies geographical data import service")
+                .version(OdreController.API_VERSION)
+                .build();
+    }
+
+    private Predicate<String> paths() {
+        return Predicates.and(PathSelectors.regex("/" + OdreController.API_VERSION + ".*"),
+                Predicates.not(PathSelectors.regex("/error.*")::apply)::apply);
+    }
+}
