+++ src/test/java/org/gridsuite/odre/server/client/BaseVoltageTest.java
@@ -1,0 +1,27 @@
+package org.gridsuite.odre.server.client;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class BaseVoltageTest {
+
+    @Test
+    public void test() {
+        BaseVoltage baseVoltage = BaseVoltage.VL_90_KV;
+        assertEquals(baseVoltage.getValue(), 90);
+
+        assertEquals(400, BaseVoltage.parse("400KV").getValue());
+        assertEquals(225, BaseVoltage.parse("225KV").getValue());
+        assertEquals(150, BaseVoltage.parse("150KV").getValue());
+        assertEquals(90, BaseVoltage.parse("90KV").getValue());
+        assertEquals(63, BaseVoltage.parse("63KV").getValue());
+        assertEquals(45, BaseVoltage.parse("45KV").getValue());
+        assertEquals(-1, BaseVoltage.parse("INFERIEUR A 45kV").getValue());
+        assertEquals(-1, BaseVoltage.parse("INF 45kV").getValue());
+        assertEquals(-1, BaseVoltage.parse("<45kV").getValue());
+        assertEquals(0, BaseVoltage.parse("HORS TENSION").getValue());
+        assertEquals(-99, BaseVoltage.parse("COURANT CONTINU").getValue());
+
+    }
+}
